Update state management for Home page
